+++ twitter4j-v2-support/src/main/kotlin/twitter4j/ListsEx.kt
@@ -144,6 +144,57 @@ fun Twitter.getListTweets(
 }
 
 /**
+ * Returns a list of members from a specified List
+ *
+ * @throws TwitterException when Twitter service or network is unavailable
+ * @see "https://developer.twitter.com/en/docs/twitter-api/lists/list-members/api-reference/get-lists-id-members"
+ */
+@Throws(TwitterException::class)
+fun Twitter.getListMembers(
+    /**
+     * The ID of the List whose members you would like to retrieve.
+     */
+    id: Long,
+    expansions: String? = null,
+    maxResults: Int? = null,
+    paginationToken: String? = null,
+    tweetFields: String? = null,
+    userFields: String? = null,
+): UsersResponse {
+
+    if (this !is TwitterImpl) throw IllegalStateException("invalid twitter4j impl")
+
+    ensureAuthorizationEnabled()
+
+    val params = ArrayList<HttpParameter>()
+
+    if (expansions != null) {
+        params.add(HttpParameter("expansions", expansions))
+    }
+
+    if (maxResults != null) {
+        params.add(HttpParameter("max_results", maxResults))
+    }
+
+    if (paginationToken != null) {
+        params.add(HttpParameter("pagination_token", paginationToken))
+    }
+
+    if (tweetFields != null) {
+        params.add(HttpParameter("tweet.fields", tweetFields))
+    }
+
+    if (userFields != null) {
+        params.add(HttpParameter("user.fields", userFields))
+    }
+
+    return V2ResponseFactory().createUsersResponse(
+        http.get(conf.v2Configuration.baseURL + "lists/" + id + "/members", params.toTypedArray(), auth, this),
+        conf
+    )
+}
+
+/**
  * Enables the authenticated user to create a List.
  *
  * @throws TwitterException when Twitter service or network is unavailable