+++ CHANGELOG.md
@@ -8,6 +8,8 @@ v1.0.0 (2021.11.xx)
   - Add Twitter.getOwnedLists() for "GET /2/users/:id/owned_lists"
 - Lists Tweets lookup
   - Add Twitter.getListTweets() for "GET /2/lists/:id/tweets"
+- List members
+  - Add Twitter.getListMembers() for "GET /2/lists/:id/members"
 - Bump Kotlin 1.6.0
 
 v0.4.0 (2021.11.12)